+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 ### 0.9.3
+- **Add**: `SessionManager.request` method, that takes an array as a request parameter.
+- **Add**: `ParametersType` enum to pass array parameter to request body.
+
+### 0.9.3
 - **Add**: `Insert`/`Remove` section with animation functions to `TableKit`. Also make new function `Replace` that uses new `Insert`/`Remove` to animate section replace.
 
 ### 0.9.2